+++ flap/src/main/java/me/yifeiyuan/flap/IFlap.java
@@ -7,6 +7,6 @@ package me.yifeiyuan.flap;
  * @author 程序亦非猿 [Follow me](<a> https://github.com/AlanCheen</a>)
  * @version 1.0
  */
-interface IFlap extends ItemFactoryManager, FlapAdapterDelegate {
+interface IFlap extends ItemFactoryManager, AdapterDelegate {
 
 }
